Updated UI Issues and Gizmos
Completed the USS Selectors and Icons. The icons are in a separate package and will need to be transferred over to the repository.

Modified the Description to be a function that is called when the inspector is updated. This allows for instant changes of text while modifying the variables.

Changes Gizmos from Traversal to a for loop. This allow for Trigger Nodes and their children to have Gizmos drawn.