Added Event Activators and Listeners, Errors, OnDisable, etc.
Event Activators and Listeners have been added,

Error states have been added and along with the corresponding selectors.

OnDisable was added to each of the nodes as a virtual function. This allows for listeners to be removed from events.

50+ nodes have been added for the following sections:
Math
String
Global Events
UI

More Infomration will be in the complete patch Notes.