Implemented String Manipulation and Node Property Changes
String Manipulation allows for any variable as long as it is not null to provide a ToString(); without knowing to the type and only using the PropertyNode.

Node Property now allows for an attribute to be used to only allow for that type of PropertyNode to be used. This is useful with nodes that have to be overridden or have objects taken from the scene.