Updated Search Window
Search window now allows for folder structures within the menuPath. This places all corresponding nodes in submenus for easier navigation and clarity. Add several new nodes and isolated a graphical bug, this bug can be squashed at a later date due to the changes that will have to be made to the description and error handling.