Added Icon Visibility Settings
Settings were added to both Icon and Description in the behaviour tree settings. This allows for the nodes to have icons or descriptions active by default. The toggles within the editor are for runtime only and are not saved to the settings file.

Fixed a couple of graphical issues with the description and errors.

Changed a couple of node icons and descriptions.