Added Raycasting Node and Rotation
Modified the GameObject Set Active node to allow for self to be turned off.

Raycasting works by being called as an action node. If the ray is successful it saves the collider, gameobject and point the raycast hit.

Rotate over time is the first of the Transform nodes that allow for simple transform functionality to be used.