Updated NumericProperty to be Simpler
Numeric Property (Number) key will be now used for all mathematics and comparison nodes as it will allow for all types: double, float, int, & long; to be used. This should simplify the number of nodes needed for basic tasks.

Another Thought is to add a similar type to numericProperty but as a vector. This vector will include all the different vectors and convert data over if needed.